fix(scaffold,docs): the sources tuple moves to @useatlas/schemas, and the connector guide learns scope (#5203)
Two CI reds, one cause each:

- Scaffold (docker/vercel): the template installs @useatlas/types from the
  registry, so the new KNOWLEDGE_COLLECTION_SOURCES value export was missing
  at scaffold build — the exact publish-first gotcha admin-schemas.ts and
  schemas/brain.ts already document. The tuple now lives in
  @useatlas/schemas (vendored into templates as source), with the
  _BrainFactStatusesCovered-shaped exhaustiveness pin; @useatlas/types keeps
  a hand-written union and a pointer comment.

- drift: BrainSourceConnector gained a required scope member and the
  published snippet in brain-connector-authoring.mdx did not. The guide now
  carries the member, the per-install/per-workspace contract, the chat-class
  registration refusal, and a checklist line.
