+++ packages/api/src/api/routes/admin-knowledge.ts
@@ -28,7 +28,6 @@
 
 import { createRoute, z } from "@hono/zod-openapi";
 import {
-  KNOWLEDGE_COLLECTION_SOURCES,
   type KnowledgeCollectionListResponse,
   type KnowledgeCollectionSource,
   type KnowledgeDocumentListResponse,
@@ -36,6 +35,7 @@ import {
   type KnowledgeSyncRunResponse,
   type KnowledgeUninstallResponse,
 } from "@useatlas/types";
+import { KNOWLEDGE_COLLECTION_SOURCES } from "@useatlas/schemas";
 import { createLogger } from "@atlas/api/lib/logger";
 import { logAdminAction, logAdminActionAwait, ADMIN_ACTIONS } from "@atlas/api/lib/audit";
 import { internalQuery } from "@atlas/api/lib/db/internal";
@@ -213,8 +213,8 @@ const CollectionListResponseSchema = z.object({
   collections: z.array(
     z.object({
       slug: z.string(),
-      // Fed from the wire type's own values array — #5203 had to remove a
-      // member from three hand-synchronized spellings of this set.
+      // Fed from the shared tuple in @useatlas/schemas — #5203 had to remove
+      // a member from three hand-synchronized spellings of this set.
       source: z.enum(KNOWLEDGE_COLLECTION_SOURCES),
       description: z.string().nullable(),
       installedAt: z.string().nullable(),
